Changement de la methode main et simplification avec une classe Jeu
